Python-01-URL-Shortner/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
WTForms==2.3.3
SQLAlchemy==1.4.23
python-dotenv==0.19.0
pytest==6.2.5
# mysqlclient==2.1.0
